add support for floats

+++ PY101-109/calculator.py
@@ -20,7 +20,7 @@ def prompt(message):
 def invalid_number(number_str):
     ''' check if numeric input is a valid integer'''
     try:
-        int(number_str)
+        float(number_str)
     except ValueError:
         return True
 
@@ -56,13 +56,13 @@ while keep_calculating:
     match operation:
 
         case '1':   # '1' represents addition
-            output = int(number1) + int(number2)
+            output = float(number1) + float(number2)
         case '2': # '2' represents subtraction
-            output = int(number1) - int(number2)
+            output = float(number1) - float(number2)
         case '3': # '3' represents multiplication
-            output = int(number1) * int(number2)
+            output = float(number1) * float(number2)
         case '4': # '4' represents division
-            output = int(number1) / int(number2)
+            output = float(number1) / float(number2)
 
     print(f"{configs['result']}{output}")
